перенес логику на уровень сервисов
