+++ GameController/GameController.API/Mappers/AuthMapper.cs
@@ -9,7 +9,15 @@ namespace GameController.API.Mappers
         public AuthMapper()
         {
             CreateMap<LoginRequest, LoginDto>();
+            CreateMap<TokenRequest, TokenDto>();
             CreateMap<TokenResponseDto, TokenResponse>();
+            CreateMap<TokenResponseDto, TokenRequest>()
+           .ForMember(
+                x => x.Token,
+                opt => opt.MapFrom(src => src.AccessToken))
+           .ForMember(
+                x => x.RefreshToken,
+                opt => opt.MapFrom(src => src.RefreshToken));
         }
     }
 }
