+++ GameController/GameController.Services/Interfaces/IAuthService.cs
@@ -1,0 +1,20 @@
+﻿using GameController.Services.Models.Auth;
+
+namespace GameController.Services.Interfaces
+{
+    /// <summary>
+    /// Interface for working with authentication.
+    /// </summary>
+    public interface IAuthService
+    {
+        /// <summary>
+        /// Validates the user credentials.
+        /// </summary>
+        TokenResponseDto ValidateUser(LoginDto loginDto);
+
+        /// <summary>
+        /// Retrieving refresh token.
+        /// </summary>
+        TokenResponseDto RefreshToken(TokenDto tokenModel);
+    }
+}
